--- src/stories/YesNoButttons/YesNoButtons.css
@@ -1,12 +1,0 @@
-/* YesNoButtons.css */
-
-button.yes.selected {
-    background-color: #2196F3; /* Blue color for selected "Yes" */
-    color: white;
-  }
-  
-  button.no.selected {
-    background-color: #FF5733; /* Red color for selected "No" */
-    color: white;
-  }
-
